Refactor Service and DesktopStarter to have the same base.
